+++ CHANGELOG.md
@@ -19,6 +19,18 @@
 
 </details>
 
+## 8.71.44 (Jul 12, 2022)
+
+## :bug: Bug Fix
+
+- `ricos-editor`
+  - [#4167](https://github.com/wix/ricos/pull/4167) fix toolbars should create (mobile)
+
+## :house: Internal
+
+- `video`
+  - [#4077](https://github.com/wix/ricos/pull/4077) add tiptap settings button
+
 ## 8.71.43 (Jul 11, 2022)
 
 - `link`